Fixed prototype of strerror (size_t --> int)
git-svn-id: https://svn.aros.org/svn/aros/trunk@23088 fb15a70f-31f2-0310-bbcc-cdcc74a49acc

+++ compiler/clib/include/string.h
@@ -24,7 +24,7 @@ char * strdup (const char * str);
 char * strcpy (char * restrict dest, const char * restrict src);
 char * strcat (char * restrict dest, const char * restrict src);
 size_t strcspn (const char *s1, const char *s2);
-char * strerror (size_t n);
+char * strerror (int n);
 char * strncat (char * restrict dest, const char * restrict src, size_t n);
 char * strncpy (char * restrict dest, const char * restrict src, size_t n);
 char * strpbrk (const char * s1, const char * s2);
